refactor: extract desktop settings adapter
